/clubs/list page built
