+++ less/components/page.less
@@ -70,14 +70,4 @@ a.skip-to-content:focus {
       }
     }
   }
-  &.clubs {
-    main {
-      ul {
-        .listItemBorderColorizer(@orangeSidebarColor);
-      }
-    }
-    .btn[type="submit"] {
-      .submitBtn(@orangeSidebarColor);
-    }
-  }
 }
