+++ less/components/sidebar.less
@@ -24,7 +24,7 @@ html.no-js .sidebar {
     .sidebarColorizer(@greenSidebarColor);
   }
   &.clubs,
-  &.clubs-list,
+  &.clubs-list-page,
   &.clubs-toolkit {
     .sidebarColorizer(@orangeSidebarColor);
   }
